+++ radiant/dags/import_kf_variants.py
@@ -48,7 +48,6 @@ with DAG(
     insert_kf_variant = StarRocksSQLExecuteQueryOperator(
         task_id="insert",
         sql="./sql/kf/kf_variants_insert.sql",
-        submit_task=True,
         submit_task_options=std_submit_task_opts,
     )
 
@@ -77,7 +76,6 @@ with DAG(
         insert_part = StarRocksSQLExecuteQueryOperator.partial(
             task_id="insert",
             sql="./sql/kf/kf_variants_part_insert_part.sql",
-            submit_task=True,
             submit_task_options=std_submit_task_opts,
             pool="starrocks_insert_pool",
             pool_slots=1,
@@ -97,7 +95,6 @@ with DAG(
         overwrite_part = StarRocksSQLExecuteQueryOperator.partial(
             task_id="overwrite",
             sql="./sql/kf/kf_variants_part_overwrite_part.sql",
-            submit_task=True,
             submit_task_options=std_submit_task_opts,
             pool="starrocks_insert_pool",
             pool_slots=1,
