+++ tests/integration/dags/test_import_kf_open_data.py
@@ -1,8 +1,13 @@
-import time
+import logging
 
-from tests.utils.dag_run import poll_dag_until_success
+import pytest
+
+from tests.utils.dags import poll_dag_until_success
+
+LOGGER = logging.getLogger(__name__)
 
 
+@pytest.mark.slow
 def test_open_data_iceberg_tables(
     init_iceberg_tables, starrocks_iceberg_catalog, setup_namespace, radiant_airflow_container, starrocks_session
 ):
@@ -13,29 +18,20 @@ def test_open_data_iceberg_tables(
     dag_id = "radiant-import-kf-open-data"
     _conf = f'{{"iceberg_catalog": "{starrocks_iceberg_catalog.name}", "iceberg_database": "{setup_namespace}"}}'
 
-    time.sleep(30)
-
     # Hashes are required to be imported first, but standalone airflow doesn't support
     # running nested dags, therefore we need to trigger them sequentially manually
     radiant_airflow_container.exec(["airflow", "dags", "unpause", hashes_dag_id])
     radiant_airflow_container.exec(["airflow", "dags", "trigger", hashes_dag_id, "--conf", _conf])
 
-    assert poll_dag_until_success(airflow_container=radiant_airflow_container, dag_id=hashes_dag_id, timeout=120)
+    assert poll_dag_until_success(airflow_container=radiant_airflow_container, dag_id=hashes_dag_id, timeout=180)
 
     radiant_airflow_container.exec(["airflow", "dags", "unpause", dag_id])
     radiant_airflow_container.exec(["airflow", "dags", "trigger", dag_id, "--conf", _conf])
 
-    assert poll_dag_until_success(airflow_container=radiant_airflow_container, dag_id=dag_id, timeout=120)
+    assert poll_dag_until_success(airflow_container=radiant_airflow_container, dag_id=dag_id, timeout=180)
 
     with starrocks_session.cursor() as cursor:
-        for _table in [
-            "1000_genomes",
-            "clinvar",
-            "dbnsfp",
-            "gnomad_genomes_v3",
-            "spliceai",
-            "topmed_bravo",
-        ]:
+        for _table in ["1000_genomes", "clinvar", "dbnsfp", "gnomad_genomes_v3", "spliceai", "topmed_bravo"]:
             cursor.execute(f"SELECT COUNT(1) FROM {_table}")
             response = cursor.fetchall()
             assert response[0][0] == 100, f"Table {_table} is empty"
