+++ tests/integration/test_starrocks.py
@@ -8,7 +8,7 @@ def test_starrocks_iceberg_state(starrocks_iceberg_catalog, setup_namespace, sta
     with starrocks_session.cursor() as cursor:
         cursor.execute(f"SHOW TABLES FROM {starrocks_iceberg_catalog.name}.{setup_namespace}")
         iceberg_tables = cursor.fetchall()
-        assert len(iceberg_tables) == 9
+        assert len(iceberg_tables) == 3
 
 
 def test_starrocks_create_table(starrocks_session):
